refactor(storybook-radiant-vite): share CSF story compose for SSR and hydrate
